第三版BlhClient, 重构所有变量
第三版BlhClient, 重构所有变量

+++ BlhClient.py
@@ -4,21 +4,23 @@ from threading  import Thread
 from queue import Queue, Empty
 import os
 import time
-from pathlib import *
+from pathlib import Path
 import urllib3
+from utils.stext import *
 
+Prefix = '!!blh'
 datahead = '§cBlhControlThread§r/§e{0} §r'
 cmd = '!!blh'
 rooturl = 'bs.s1.blackserver.cn:8900'
-selfname = ''
+stopName = ''
 ON_POSIX = 'posix' in sys.builtin_module_names
-p = Popen('', shell = True)
+popu = []
+roomnames = []
 stopflag = True
 ud = False
-t = None
-q = None
 
 helpmsg = '''§e-------------WebSocketBlhClient--help-------------
+{0} cmds: §b命令交互列表(ClickEvent)
 {0}: §b显示help
 {0} help [页数(可选,默认1)]: §b显示help
 {0} add [自定义名称] [房间id]: §b向配置文件里添加一个blh启动项
@@ -26,15 +28,70 @@ helpmsg = '''§e-------------WebSocketBlhClient--help-------------
 {0} start [名称]: §b启动一个在配置文件blhClient
 {0} stop [名称]: §b停止一个正在运行的blhClient
 {0} stop all: §b停止所有正在运行的blhClient
-{0} list: §b列出所有在配置文件中的可用直播间
 §e-------------------页2/1--------------------'''.format('§n§7' + cmd)
 
 helpmsg_2 = '''§e-------------WebSocketBlhClient--help-------------
+{0} pop [名称]: §b关闭/开启 一个已经启动的blh的人气
+{0} list: §b列出所有在配置文件中的可用直播间
+{0} listrun: §b列出所有正在运行的直播间名称
 {0} v: §b查看现在运行的BlhClient版本
 {0} vmsg: §b查看现在运行的BlhClient更新消息
 {0} reload: §b重载此插件的内容(一般用不到)
 {0} checkud: §b检查版本更新
 §e-------------------页2/2------------------------'''.format('§n§7' + cmd)
+
+def get_text(t1, t2='', t3='', color=SColor.white, run=True):
+    if t2 != '':
+        if t3 != '':
+            if run:
+                stxt = SText(t1, color=color).set_hover_text(
+                    t2).set_click_event(SAction.run_command, t3)
+            else:
+                stxt = SText(t1, color=color).set_hover_text(
+                    t2).set_click_event(SAction.suggest_command, t3)
+        else:
+            stxt = SText(t1, color=color,
+                         styles=SStyle.italic).set_hover_text(t2)
+    else:
+        if t3 != '':
+            stxt = SText(t1, color=color)
+        else:
+            if run:
+                stxt = SText(t1, color=color).set_click_event(
+                    SAction.run_command, t3)
+            else:
+                stxt = SText(t1, color=color).set_click_event(
+                    SAction.suggest_command, t3)
+    return stxt
+
+
+easycmds = STextList(
+    get_text('§e------------§bBlhClient§e------------\n   '),
+    get_text(f'§7{Prefix} add [名称] [roomid]', '§b添加一个房间', f'{Prefix} add ', run=False),
+    get_text(' §r§b添加房间\n   '),
+    get_text(f'§7{Prefix} rm [名称]', '§b用名称删除房间', f'{Prefix} rm ',run=False),
+    get_text(' §r§b删除房间\n   '),
+    get_text(f'§7{Prefix} start [名称]', '§b启动一个在配置文件里的blh', f'{Prefix} start ', run=False),
+    get_text(' §r§b订阅房间\n   '),
+    get_text(f'§7{Prefix} stop [名称]', '§b停止一个正在运行的blh', f'{Prefix} stop ',run=False),
+    get_text(' §r§b取消订阅\n   '),
+    get_text(f'§7{Prefix} stop all', '§b停止所有', f'{Prefix} stop all'),
+    get_text(' §r§b停止所有\n   '),
+    get_text(f'§7{Prefix} pop [名称]', '§b开/关 一个正在运行的blh人气显示', f'{Prefix} pop ', run=False),
+    get_text(' §r§b人气开关\n'),
+    get_text(f'§7   {Prefix} listrun','§b列出正在运行的直播间', f'{Prefix} listrun'),
+    get_text(' §r§b列表\n'),
+    get_text(f'§7   {Prefix} v', '§b查看现在运行的BlhClient版本', f'{Prefix} v'),
+    get_text(' §r§b版本查看\n'),
+    get_text(f'§7   {Prefix} vmsg', '§b查看现在运行的BlhClient更新消息', f'{Prefix} vmsg'),
+    get_text(' §r§b版本信息查看\n'),
+    get_text(f'§7   {Prefix} reload', '§b重载blh', f'{Prefix} reload'),
+    get_text(' §r§b重载blh\n'),
+    get_text(f'§7   {Prefix} checkud', '§b检查版本更新', f'{Prefix} reload'),
+    get_text(' §r§b更新blh'),
+)
+
+
 
 def printHelp(server, player, msg):
     for line in msg.splitlines():
@@ -87,8 +144,16 @@ def remove(key):
 
 
 
-def start_dm(server, roomid):
-    global stopflag, p, selfname, q, t
+def blh(server, info, args):
+    global stopflag, stopName
+    roomid = load(args[2])
+    if roomid is None:
+        dm_logger_tell(server, '§cBlhControlThread§r/§e{0}§r §a key not found'.format('info'), info.player)
+        return
+    if test_blh_isrunning(server, info.player, args[2]) is True:
+        return
+    test_file(server, info.player)
+    selfname = args[2]
     stopflag = True
     p = Popen('python3 demo.py {0} True'.format(roomid), stdout = PIPE, shell = True, bufsize = 1, close_fds = ON_POSIX, cwd = 'plugins/blh')
     dm_logger(server, '§cBlhControlThread§r/§e{0}§r §a blh is running at pid {1}'.format('info', str(p.pid)))
@@ -96,25 +161,37 @@ def start_dm(server, roomid):
     t = Thread(target=enqueue_output, args=(p.stdout, q))
     t.daemon = True
     t.start()
-    dm_logger(server, datahead.format('info') + '已订阅{}的直播间'.format(selfname))
-    time.sleep(2)
-
-def stop_dm(server, player):
-    global stopflag, p
-    code = 0
-    stopflag = False
-    os.system('kill {}'.format(str(p.pid)))
-    p.kill()
-    while p.poll() == None:
-        os.system('kill {}'.format(str(p.pid)))
-        p.kill()
-        dm_logger_tell(server, '§cBlhControlThread§r/§e{}§r§astopping'.format('info '), player)
-        code = p.poll()
-    dm_logger_tell(server, '§cBlhControlThread§r/§e{0}§r§aexitcode {1}'.format('info ', str(code)), player)
+    dm_logger(server, datahead.format('info') + '已订阅{0}的直播间'.format(selfname))
+    while stopflag and p.poll() is None:
+            if selfname == stopName:
+                roomnames.remove(selfname)
+                tmp = p.pid
+                p.kill()
+                os.system('kill {0}'.format(str(tmp)))
+                t._delete()
+                stopName = ''
+                return
+            try:
+                roomnames.index(selfname)
+            except ValueError:
+                break
+            try:
+                line = q.get_nowait()
+            except Empty:
+                pass
+            else:
+                try:
+                    popu.index(selfname)
+                except ValueError:
+                    dm_logger(server, line, selfname)
+                else:
+                    if bytes('当前人气值|', encoding='utf-8') not in line:
+                        dm_logger(server, line, selfname)
 
 
 
 def dm_logger(server, data, name = "defalt"):
+    global popu
     try:
         data = str(data, encoding='utf-8')
     except:
@@ -138,9 +215,6 @@ def dm_logger(server, data, name = "defalt"):
         args = buff.split('|')
         server.say('§b[§cBLH§r§b][§c{0}§b] §c{1}§r:{2}'.format(name, args[0], args[1]))
     else:
-        if '人气' in buff:
-            server.say('111')
-            return None
         server.say('§b[§cBLH§r§b]§r ' + buff)
 
 def dm_logger_tell(server, buff, player = '@a'):
@@ -149,13 +223,15 @@ def dm_logger_tell(server, buff, player = '@a'):
 
 
 def test_blh_isrunning(server, player, name):
-    global stopflag, p, selfname
-    if name == selfname:
-        dm_logger_tell(server, '§cBlhControlThread§r/§4§l{0}§r §athis blh:{1} is running'.format('Warn', selfname), player)
-        return True
-    else:
-        selfname = name
+    global stopflag, roomnames
+    try:
+        roomnames.index(name)
+    except ValueError:
+        roomnames.append(name)
         return False
+    else:
+        dm_logger_tell(server, '§cBlhControlThread§r/§4§l{0}§r §athis blh:{1} is running'.format('Warn', name), player)
+        return True
 
 def test_file(server, player):
     filedir = Path('plugins/blh')
@@ -220,6 +296,7 @@ def test_file(server, player):
             pass
 
 def check_update(server, player):
+    global roomnames, stopflag
     dm_logger(server, '检查更新中,请耐心等待.期间请勿重载插件!')
     url = rooturl + '/ver'
     http = urllib3.PoolManager()
@@ -227,45 +304,49 @@ def check_update(server, player):
     f = open('plugins/blh/ver', 'r')
     ver = f.read()
     if int(res.data) > int(ver):
-        dm_logger(server, '有新版本的Blh可用!')
-        dm_logger(server, '最新版本为: {0}.0.0'.format(str(res.data, encoding='utf-8')))
-        time.sleep(1.5)
-        dm_logger(server, datahead.format('info') + '停止所有房间')
-        stop_dm(server, player)
-        dm_logger(server, '等待3秒之后将开始更新.期间请勿重载插件!')
-        time.sleep(3)
-        dm_logger_tell(server, '更新: demo.py')
-        url = rooturl + '/demo.py'
-        http = urllib3.PoolManager()
-        res = http.request('GET', url)
-        os.remove('plugins/blh/demo.py')
-        f = open('plugins/blh/demo.py', 'w')
-        f.write(str(res.data, encoding='utf-8'))
-        dm_logger_tell(server, '更新: blivedm.py')
-        url = rooturl + '/blivedm.py'
-        http = urllib3.PoolManager()
-        res = http.request('GET', url)
-        os.remove('plugins/blh/blivedm.py')
-        f = open('plugins/blh/blivedm.py', 'w')
-        f.write(str(res.data, encoding='utf-8'))
-        dm_logger_tell(server, '更新版本文件')
-        url = rooturl + '/ver'
-        http = urllib3.PoolManager()
-        res = http.request('GET', url)
-        os.remove('plugins/blh/ver')
-        f = open('plugins/blh/ver', 'w')
-        f.write(str(res.data, encoding='utf-8'))
-        dm_logger_tell(server, '更新Blh主程序, Blh即将退出!')
-        url = rooturl + '/BlhClient.py'
-        http = urllib3.PoolManager()
-        res = http.request('GET', url)
-        os.remove('plugins/BlhClient.py')
-        f = open('plugins/BlhClient.py', 'w')
-        f.write(str(res.data, encoding='utf-8'))
-        dm_logger(server, datahead.format('info') + '即将重载!')
-        global ud
-        ud = True
-        server.load_plugin('BlhClient.py')
+        try:
+            dm_logger(server, '有新版本的Blh可用!')
+            dm_logger(server, '最新版本为: {0}.0.0'.format(str(res.data, encoding='utf-8')))
+            time.sleep(1.5)
+            dm_logger(server, datahead.format('info') + '停止所有房间')
+            roomnames = []
+            stopflag = False
+            dm_logger(server, '等待3秒之后将开始更新.期间请勿重载插件!')
+            time.sleep(3)
+            dm_logger_tell(server, '更新: demo.py')
+            url = rooturl + '/demo.py'
+            http = urllib3.PoolManager()
+            res = http.request('GET', url)
+            os.remove('plugins/blh/demo.py')
+            f = open('plugins/blh/demo.py', 'w')
+            f.write(str(res.data, encoding='utf-8'))
+            dm_logger_tell(server, '更新: blivedm.py')
+            url = rooturl + '/blivedm.py'
+            http = urllib3.PoolManager()
+            res = http.request('GET', url)
+            os.remove('plugins/blh/blivedm.py')
+            f = open('plugins/blh/blivedm.py', 'w')
+            f.write(str(res.data, encoding='utf-8'))
+            dm_logger_tell(server, '更新版本文件')
+            url = rooturl + '/ver'
+            http = urllib3.PoolManager()
+            res = http.request('GET', url)
+            os.remove('plugins/blh/ver')
+            f = open('plugins/blh/ver', 'w')
+            f.write(str(res.data, encoding='utf-8'))
+            dm_logger_tell(server, '更新Blh主程序, Blh即将退出!')
+            url = rooturl + '/BlhClient.py'
+            http = urllib3.PoolManager()
+            res = http.request('GET', url)
+            os.remove('plugins/BlhClient.py')
+            f = open('plugins/BlhClient.py', 'w')
+            f.write(str(res.data, encoding='utf-8'))
+            dm_logger(server, datahead.format('info') + '即将重载!')
+            global ud
+            ud = True
+            server.load_plugin('BlhClient.py')
+        except:
+            dm_logger(server, datahead.format('Err') + '检测更新: 失败(请检查网络)')
     else:
         dm_logger(server, 'blh已经是最新版本!')
         dm_logger(server, '目前版本: v{0}.0.0'.format(str(res.data, encoding='utf-8')))
@@ -276,28 +357,35 @@ def check_update(server, player):
 
 
 
-def init(server, info, startargs):
-    roomid = load(startargs[2])
-    if roomid is None:
-        dm_logger_tell(server, '§cBlhControlThread§r/§e{0}§r §a key not found'.format('info'), info.player)
-        return
-    if test_blh_isrunning(server, info.player, startargs[2]) is True:
-        return
-    test_file(server, info.player)
-    start_dm(server, roomid)
-
-    
-
 def on_server_stop(server):
-    stop_dm(server, '@a')
+    global stopflag, popu, roomnames
+    roomnames = []
+    stopflag = False
+    try:
+        os.system('ps aux|grep "python3 demo.py"|grep -v grep|cut -c 9-15|xargs kill -15')
+    except:
+        pass
 
 def on_mcdr_stop(server):
-    stop_dm(server, '@a')
+    global stopflag, popu, roomnames
+    roomnames = []
+    stopflag = False
+    try:
+        os.system('ps aux|grep "python3 demo.py"|grep -v grep|cut -c 9-15|xargs kill -15')
+    except:
+        pass
 
 def on_unload(server):
-    stop_dm(server, '@a')
+    global stopflag, popu, roomnames
+    roomnames = []
+    stopflag = False
+    try:
+        os.system('ps aux|grep "python3 demo.py"|grep -v grep|cut -c 9-15|xargs kill -15')
+    except:
+        pass
 
 def on_load(server, old):
+    global stopflag, popu, roomnames
     server.add_help_message('!!blh', 'BiliBili弹幕姬')
     test_file(server, '@a')
     os.system('chmod 777 plugins/BlhClient.py')
@@ -334,7 +422,7 @@ def on_load(server, old):
 
 
 def on_info(server, info):
-    global stopflag, p, selfname, q, t
+    global stopflag, popu, roomnames, stopName
     if info.content.startswith('!!blh'):
         startargs = info.content.split(' ')
         if len(startargs) == 1:
@@ -345,7 +433,7 @@ def on_info(server, info):
             if startargs[2] == '1':
                 printHelp(server, info.player, helpmsg)
             elif startargs[2] == '2':
-                server.tell(info.player, '\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n')
+                server.tell(info.player, '\n')
                 printHelp(server, info.player, helpmsg_2)
             else:
                 printHelp(server, info.player, helpmsg)
@@ -369,40 +457,44 @@ def on_info(server, info):
                     tmp[1] = tmp[1].replace('\n', '')
                     dm_logger_tell(server, '名称:{0} 房号:{1}'.format(tmp[0], tmp[1]))
         elif len(startargs) == 3 and startargs[1] == 'start':
-            init(server, info, startargs)
-            while stopflag and p.poll() is None:
-                try:
-                    line = q.get_nowait()
-                except Empty:
-                    pass
-                else:
-                    dm_logger(server, line, selfname)
+            blh(server, info, startargs)
         elif len(startargs) == 3 and startargs[1] == 'stop':
             if startargs[2] == 'all':
-                stop_dm(server, 'None')
-                if selfname != '':
-                    dm_logger(server, '已取消订阅{0}房间'.format('所有'))
-                    selfname = ""
+                if roomnames != []:
+                    stopflag = False
+                    stopflag = True
+                    dm_logger(server, '已取消订阅全部房间,名称:')
+                    for line in roomnames:
+                        dm_logger(server, line)
+                    roomnames = []
                     try:
                         os.system('ps aux|grep "python3 demo.py"|grep -v grep|cut -c 9-15|xargs kill -15')
                     except:
                         pass
                 else:
                     dm_logger(server, datahead.format('Warn') + '没有房间可以被取消订阅')
-            elif startargs[2] == selfname:
-                stop_dm(server, info.player)
-                dm_logger(server, '已取消订阅{0}的直播间'.format(selfname))
-                selfname = ""
+            elif startargs[2] != 'all':
+                try:
+                    roomnames.index(startargs[2])
+                except ValueError:
+                    dm_logger(server, '没有名叫: {0}的直播间'.format(startargs[2]))
+                    return
+                else:
+                    stopName = startargs[2]
+                    dm_logger(server, '已取消订阅{0}的直播间'.format(startargs[2]))
             else:
-                dm_logger_tell(server, '§cBlhControlThread§r/§e{0}§r §a 参数错误'.format('info'), info.player)
+                dm_logger(server, '参数错误!')
         elif len(startargs) == 2 and startargs[1] == 'reload':
             dm_logger_tell(server, '正在停止所有房间')
-            stop_dm(server, 'None')
-            selfname = ""
-            try:
-                os.system('ps aux|grep "python3 demo.py"|grep -v grep|cut -c 9-15|xargs kill -15')
-            except:
-                pass
+            if roomnames != []:
+                stopflag = False
+                stopflag = True
+                dm_logger(server, '已取消订阅{0}房间'.format('所有'))
+                roomnames = []
+                try:
+                    os.system('ps aux|grep "python3 demo.py"|grep -v grep|cut -c 9-15|xargs kill -15')
+                except:
+                    pass
             server.load_plugin("BlhClient.py")
         elif len(startargs) == 2 and startargs[1] == 'checkud':
             check_update(server, info.player)
@@ -415,5 +507,25 @@ def on_info(server, info):
             http = urllib3.PoolManager()
             res = http.request('GET', url)
             server.say(str(res.data, encoding='utf-8')) 
+        elif len(startargs) == 3 and startargs[1] == 'pop':
+            try:
+                roomnames.index(startargs[2])
+            except ValueError:
+                dm_logger_tell(server, datahead.format('Warn') + ' 没有房间名称为: {0}'.format(startargs[2]), info.player)
+                return
+            else:
+                try:
+                    popu.index(startargs[2])
+                except ValueError:
+                    popu.append(startargs[2])
+                    dm_logger_tell(server, datahead.format('info') + ' 关闭了房间{0}的人气显示'.format(startargs[2]), info.player)
+                else:
+                    popu.remove(startargs[2])
+                    dm_logger_tell(server, datahead.format('info') + ' 打开了房间{0}的人气显示'.format(startargs[2]), info.player)
+        elif len(startargs) == 2 and startargs[1] == 'listrun':
+            for line in roomnames:
+                dm_logger(server, line)
+        elif len(startargs) == 2 and startargs[1] == 'cmds':
+            server.tell(info.player, easycmds)
         else:
             dm_logger_tell(server, datahead.format('Warn') + '参数错误, 请!!blh help查看帮助')
